improve(web): 쿼리클라이언트 스트리밍 ssr
